+++ mizani/_core/dates.py
@@ -1,11 +1,13 @@
 from __future__ import annotations
 
 import math
+from collections.abc import Sized
 from datetime import datetime, timedelta, tzinfo
 from typing import TYPE_CHECKING, overload
 from zoneinfo import ZoneInfo
 
 import numpy as np
+import pandas as pd
 from dateutil.rrule import rrule
 
 from ..utils import get_timezone, isclose_abs
@@ -22,6 +24,8 @@ if TYPE_CHECKING:
         NDArrayDatetime,
         NDArrayFloat,
         SeqDatetime,
+        Timedelta,
+        TimedeltaArrayLike,
         TzInfo,
     )
 
@@ -149,6 +153,47 @@ def num_to_datetime(
     """
     tz = get_tzinfo(tz) or UTC
     return _from_ordinalf_np_vectorized(x, tz)
+
+
+# NOTE: We only deal with timedelta and pd.Timedelta
+
+
+@overload
+def timedelta_to_num(x: TimedeltaArrayLike) -> NDArrayFloat: ...
+
+
+@overload
+def timedelta_to_num(x: Timedelta) -> float: ...
+
+
+def timedelta_to_num(
+    x: TimedeltaArrayLike | Timedelta,
+) -> NDArrayFloat | float:
+    """
+    Convert any timedelta to days
+
+    This function gives us a numeric representation a timedelta that
+    we can add/subtract from the numeric representation of datetimes.
+    """
+    _x = x if (sized := isinstance(x, Sized)) else pd.Series([x])
+
+    if not len(_x):
+        return np.array([], dtype=float)
+
+    res: NDArrayFloat = np.array(
+        [td.total_seconds() / SECONDS_PER_DAY for td in _x]
+    )
+    return res if sized else res[0]
+
+
+def num_to_timedelta(x: FloatArrayLike) -> Sequence[pd.Timedelta]:
+    """
+    Convert any float array to numpy datetime64 array
+
+    Returns pd.Timedelta because they have a larger range than
+    datetime.timedelta.
+    """
+    return tuple(pd.Timedelta(days=val) for val in x)
 
 
 WIDTHS: dict[DateFrequency, Sequence[int]] = {